food_ai_system/requirements.txt
ultralytics
streamlit
opencv-python
pillow
groq
python-dotenv
transformers
torch
pillow